Patch submitted by "Jeff Peck" <jpjnk4-jdee@yahoo.com>:
Continuing to make JDEE safe for users of inner-classes, I offer this
patch to jde-import.el so if you have: 
import foo.Bar;

and in the code you reference: Bar.Baz or Bar.Baz.class or
Bar.Baz.ENUM or Bar.Baz.method(...), etc. JDEE will not insist on
adding another: 
import foo.Bar.Baz;

the same filter also recognizes if import foo.*; then no need to
import foo.Bar;

Note: I have made the default for this to be 'true' even though that
may change existing experience.  Because I believe the prior behaviour
is a bug, and so the default is to fix the bug.  As noted, if you
need/want the old behaviour in a particular case, just supply the
prefix-arg.

Also fixed a bug where the import insertion point did not go to the
next line after the last import. so if you had a comment on the last
import, jde would insert before that comment...



git-svn-id: svn://svn.code.sf.net/p/jdee/code/branches/2.4.0/jde@183 381a1f10-3442-0410-997e-ef331cd04102