Patch from jalhborn@users.sourceforge.net.
The JDE font locking would get confused when JDE mode was re-applied
to a buffer (typically when manually reverting a buffer or when the
user has auto-revert enabled). The patch adds the jde font lock setup
to the java-mode-hook (comments through this file are inconsistent as
to whether java-mode should be affected or not - if you want java-mode
to not employ jde font locking, setting jde-use-font-lock to nil
should disable it.

git-svn-id: svn://svn.code.sf.net/p/jdee/code/trunk/jdee@249 381a1f10-3442-0410-997e-ef331cd04102